msm: vidc: Fix broken debugfs creation error checks and error paths
The debugfs helper functions do not always return NULL when they fail;
instead, they can return an error number casted as a pointer, so that their
users have the option to determine the exact cause of failure.

Use the IS_ERR_OR_NULL() helper when checking for debugfs errors to fix the
error checks.

Change-Id: I7c2f5d45c6857e72ad2de200d642d018923040da
Signed-off-by: Sultanxda <sultanxda@gmail.com>