core_ctl: Disallow Power HAL to change min/max core limits
LineageOS ships a custom Power HAL that ruins custom settings, as
every time screen turns on/off, core_ctl tunables will be overwritten.

To avoid this, we check for system_server or PowerManager and just
return -EINVAL to disallow Power HAL to write to sysfs nodes. Yes,
it is more safe to add locks and use memcmp, but I don't think there
will be many races.

Signed-off-by: Wang Han <416810799@qq.com>
Signed-off-by: Shreyansh Lodha <slodha96@gmail.com>